requirements.txt
amqp==2.5.0
aniso8601==8.0.0
apache-beam==2.23.0
asn1crypto==0.24.0
astroid==2.4.1
avro-python3==1.9.2.1
beautifulsoup4==4.7.1
billiard==3.6.0.0
CacheControl==0.12.5
cachetools==3.1.1
carto==1.7.0
cartodb==0.8.1
certifi==2019.9.11
cffi==1.12.2
chardet==3.0.4
click==7.1.1
colorama==0.4.3
crcmod==1.7
cryptography==2.6.1
decorator==4.4.0
defusedxml==0.6.0
dialogflow==0.8.0
dill==0.3.1.1
Django==2.2.6
django-allauth==0.40.0
django-core==1.4.1
django-cors-headers==3.1.1
django-extensions==2.2.3
django-filter==2.2.0
django-filters==0.2.1
django-gapc-storage==0.5.1
django-rest-auth==0.9.5
django-storages==1.7.1
djangorestframework==3.7.7
djangorestframework-expiring-authtoken==0.1.4
djangorestframework-jwt==1.11.0
djangotoolbox==1.8.0
docopt==0.6.2
et-xmlfile==1.0.1
ez-setup==0.9
fastavro==0.23.6
firebase-admin==2.17.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
FormEncode==1.3.1
fsspec==0.5.1
future==0.18.2
gcloud==0.17.0
gcsfs==0.3.0
get==2019.4.13
google==2.0.2
google-api-core==1.17.0
google-api-python-client==1.7.11
google-auth==1.14.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
google-cloud==0.34.0
google-cloud-bigquery==1.11.2
google-cloud-core==0.29.0
google-cloud-datastore==1.7.3
google-cloud-firestore==1.6.2
google-cloud-logging==1.15.0
google-cloud-storage==1.14.0
google-cloud-tasks==1.3.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
GoogleAppEngineCloudStorageClient==1.9.22.1
googlemaps==3.0.2
grpc-google-iam-v1==0.12.3
grpcio==1.31.0
hdfs==2.5.8
httplib2==0.14.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jellyfish==0.7.2
Jinja2==2.11.1
jws==0.1.3
kombu==4.6.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mock==2.0.0
msgpack==0.6.1
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==1.4.4
Naked==0.1.31
numpy==1.16.3
oauth2==1.9.0.post1
oauth2client==3.0.0
oauthlib==3.1.0
openpyxl==2.6.3
pandas==0.25.0
pbr==5.5.0
Pillow==6.2.0
pkcs7==0.1.2
post==2019.4.13
protobuf==3.7.1
public==2019.4.13
pyarrow==0.17.1
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
pydot==1.4.1
PyJWT==1.7.1
pylint==2.5.2
pymongo==3.11.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrestcli==0.6.11
PySocks==1.7.1
python-dateutil==2.8.0
python-firebase==1.2
python-http-client==3.2.1
python-jwt==2.0.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.1.1
query-string==2019.4.13
request==1.0.2
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.7.0
rsa==4.0
segno==1.0.0
sendgrid==6.1.0
shellescape==3.8.1
six==1.12.0
soupsieve==1.9.1
SQLAlchemy==1.3.16
sqlparse==0.3.0
style==1.1.0
toml==0.10.1
tweepy==3.8.0
twilio==6.31.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
update==0.0.1
uritemplate==3.0.0
urllib3==1.25.6
vine==1.3.0
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
